+++ app/src/main/java/com/hujiang/restvolley/demo/RestfulAPIActivity.java
@@ -3,7 +3,10 @@ package com.hujiang.restvolley.demo;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
+import android.widget.LinearLayout;
+import android.widget.TextView;
 
+import com.hujiang.restvolley.demo.models.UserInfo;
 import com.hujiang.restvolley.webapi.RestVolleyCallback;
 import com.hujiang.restvolley.webapi.request.GetRequest;
 
@@ -13,27 +16,42 @@ public class RestfulAPIActivity extends AppCompatActivity {
 
     private static final String TAG = RestfulAPIActivity.class.getSimpleName();
 
-    private static final String[] CONCURRENCE_URLS = {"http://www.baidu.com"
+    private static final String[] CONCURRENCE_URLS = {
+            "http://www.baidu.com"
             , "http://stackoverflow.com/"
             , "http://www.hujiang.com"
             , "https://bintray.com/"
-            , "http://www.infoq.com/cn/"};
+            , "http://www.infoq.com/cn/"
+    };
 
+    private LinearLayout mRestGroup;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_restful_api);
 
-        for (String s : CONCURRENCE_URLS) {
-            new GetRequest(RestfulAPIActivity.this).url(s).execute(new RestVolleyCallback<String>() {
+        mRestGroup = (LinearLayout) findViewById(R.id.rest_group);
+
+        for (final String s : CONCURRENCE_URLS) {
+            new GetRequest(RestfulAPIActivity.this).url(s).execute(String.class, new RestVolleyCallback<String>() {
                 @Override
                 public void onSuccess(int statusCode, String data, Map<String, String> headers, boolean notModified, long networkTimeMs, String message) {
-                    Log.i(TAG, statusCode + "::" + data);
+                    StringBuilder builder = new StringBuilder();
+                    builder.append("\n+++++++++++++++++++++++++++++\n")
+                            .append(s).append("::").append(statusCode).append(":SUCCESSFUL:").append('\n');
+                    TextView textView = new TextView(RestfulAPIActivity.this);
+                    textView.setText(builder.toString());
+                    mRestGroup.addView(textView);
                 }
 
                 @Override
                 public void onFail(int statusCode, String data, Map<String, String> headers, boolean notModified, long networkTimeMs, String message) {
-                    Log.i(TAG, statusCode + "::" + message);
+                    StringBuilder builder = new StringBuilder();
+                    builder.append("\n+++++++++++++++++++++++++++++\n")
+                            .append(s).append("::").append(statusCode).append(":FAILURE:").append('\n').append(data).append("::\n").append(message);
+                    TextView textView = new TextView(RestfulAPIActivity.this);
+                    textView.setText(builder.toString());
+                    mRestGroup.addView(textView);
                 }
             });
         }