+++ CHANGELOG.md
@@ -1,5 +1,13 @@
+
 ### v1.0.12(2016-12-29)
 * 优化图片内存问题
+* 解决清图片缓存是DiskLruCache出现IllegalStateException:cache is closed
+* RestVolleyImageLoader支持流式图片下载，降低内存占用率
+* 优化https支持, 默认信任所有请求，安全要求较高的app可以通过接口设置SSLSocketFactory和HostnameVerifier，或者添加证书锁定CertificatePinner来增加安全性
+* 提供常量CertificateUtils.DEFAULT_HOSTNAME_VERIFIER; CertificateUtils.ALLOW_ALL_HOSTNAME_VERIFIER
+* 提供接口CertificateUtils.getDefaultSSLSocketFactory()
+* 支持内置自签名证书CertificateUtils.getSSLSocketFactory(InputStream certificateInputStream)
+* 优化下载逻辑，支持HEAD头不传Content-Length的下载
 
 ### v1.0.11(2016-12-28)
 * 解决RestVolleyImageLoader加载本地大图出现OOM的问题，但是网络大图下载的问题暂时没有解决，后续会重写网络图片下载那块逻辑，解决大图下载问题